vk: extract long functions from vk_image
